montañita_adopta/backend_FastAPI/requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
requests
python-multipart
passlib
python-dotenv
pyjwt
Jinja2
mysql-connector-python
sqlalchemy
